requirements.txt
numpy==1.26.4
pytesseract==0.3.10
opencv-python==4.10.0.84
pillow==10.3.0
pillow_heif==0.17.0
PyMuPDF==1.24.7
comtypes==1.4.4
reportlab==4.2.2
PyDrive==1.3.1
